+++ crates/cheatnet/tests/cheatcodes/deploy.rs
@@ -113,8 +113,7 @@ fn call_predefined_contract_from_proxy_contract() {
         "CheatCallerAddressCheckerProxy",
         &[],
     );
-    let proxy_selector =
-        felt_selector_from_name("get_cheated_caller_address");
+    let proxy_selector = felt_selector_from_name("get_cheated_caller_address");
     let output = call_contract(
         &mut cached_state,
         &mut cheatnet_state,
