+++ crates/cheatnet/tests/cheatcodes/mod.rs
@@ -4,20 +4,20 @@ use cheatnet::runtime_extensions::forge_runtime_extension::cheatcodes::storage::
 
 mod test_environment;
 
+mod cheat_block_number;
+mod cheat_block_timestamp;
+mod cheat_caller_address;
+mod cheat_sequencer_address;
 mod declare;
 mod deploy;
-mod elect;
 mod get_class_hash;
 mod load;
 mod mock_call;
-mod prank;
 mod precalculate_address;
 mod replace_bytecode;
-mod roll;
 mod spoof;
 mod spy_events;
 mod store;
-mod warp;
 
 pub fn map_entry_address(var_name: &str, key: &[Felt252]) -> Felt252 {
     calculate_variable_address(&felt_selector_from_name(var_name), Some(key))
