+++ crates/cheatnet/tests/contracts/src/cheat_caller_address.cairo
@@ -1,0 +1,5 @@
+mod checker;
+mod constructor_checker;
+mod checker_library_call;
+mod checker_proxy;
+mod checker_cairo0;
