+++ crates/cheatnet/tests/contracts/src/cheat_sequencer_address.cairo
@@ -1,0 +1,4 @@
+mod checker;
+mod constructor_checker;
+mod checker_library_call;
+mod checker_proxy;
