--- crates/cheatnet/tests/contracts/src/elect.cairo
@@ -1,4 +1,0 @@
-mod elect_checker;
-mod constructor_elect_checker;
-mod elect_checker_library_call;
-mod elect_checker_proxy;
