+++ crates/cheatnet/tests/contracts/src/lib.cairo
@@ -2,13 +2,13 @@ mod common;
 mod events;
 mod get_class_hash;
 mod mock;
-mod prank;
+mod cheat_caller_address;
 mod replace_bytecode;
-mod roll;
+mod cheat_block_number;
 mod spoof;
-mod elect;
+mod cheat_sequencer_address;
 mod starknet;
-mod warp;
+mod cheat_block_timestamp;
 mod segment_arena_user;
 mod panic_call;
 mod store_load;
