--- crates/cheatnet/tests/contracts/src/prank.cairo
@@ -1,5 +1,0 @@
-mod prank_checker;
-mod constructor_prank_checker;
-mod prank_checker_library_call;
-mod prank_checker_proxy;
-mod prank_checker_cairo0;
