--- crates/cheatnet/tests/contracts/src/roll.cairo
@@ -1,4 +1,0 @@
-mod roll_checker;
-mod constructor_roll_checker;
-mod roll_checker_library_call;
-mod roll_checker_proxy;
