--- crates/cheatnet/tests/contracts/src/warp.cairo
@@ -1,4 +1,0 @@
-mod warp_checker;
-mod constructor_warp_checker;
-mod warp_checker_library_call;
-mod warp_checker_proxy;
