+++ crates/forge/tests/integration/mod.rs
@@ -1,23 +1,24 @@
 mod available_gas;
+mod cheat_block_number;
+mod cheat_block_timestamp;
+mod cheat_caller_address;
 mod cheat_fork;
+mod cheat_sequencer_address;
 mod declare;
 mod deploy;
 mod deploy_at;
 mod dict;
 mod dispatchers;
-mod elect;
 mod env;
 mod fuzzing;
 mod gas;
 mod get_class_hash;
 mod l1_handler_executor;
 mod mock_call;
-mod prank;
 mod precalculate_address;
 mod pure_cairo;
 mod replace_bytecode;
 mod resources;
-mod roll;
 mod runtime;
 mod setup_fork;
 mod should_panic;
@@ -29,4 +30,3 @@ mod syscalls;
 mod test_state;
 mod too_many_events;
 mod trace;
-mod warp;
